fix #4858
